refactor: drop redundant `as i32` casts left by c_int removal
Replacing the c_int alias with i32 unmasked a large batch of
clippy::unnecessary_cast warnings on expressions like `0x0002 as i32`
/ `LITERAL as i32` (and equivalent `as usize`/`as isize` literal
casts). The c_int alias had previously hidden these from the lint;
once it became a plain i32 the casts were genuinely redundant.

Fixed mechanically via `cargo clippy --fix` scoped to only
unnecessary_cast (`-A clippy::all -W clippy::unnecessary_cast`),
followed by `cargo fmt`. The diff is exclusively removal of redundant
literal casts (clippy rewrites `1 as i32` to the typed literal `1_i32`
etc.); no other lint or behavior changes.

cargo build clean (0 warnings); cargo test --lib 257 passed;
cargo fmt --check clean; `cargo clippy --lib` now 301 warnings, back
under the origin/main baseline of 332.

+++ src/ar.rs
@@ -139,7 +139,7 @@ pub unsafe fn ar_parse_name(
         );
     }
     let fresh0 = p;
-    p = p.offset(1 as i32 as isize);
+    p = p.offset(1_i32 as isize);
     *fresh0 = 0;
     *p.offset(strlen(p).wrapping_sub(1) as isize) = 0;
     *memname_p = p;
@@ -251,13 +251,13 @@ pub unsafe fn ar_member_date(name: *const ::core::ffi::c_char) -> time_t {
     );
     if (0 as intmax_t) < val
         && val
-            <= (if (0 as i32 as time_t) < -(1 as i32) as time_t {
-                -(1 as i32) as time_t
+            <= (if (0_i32 as time_t) < -1_i32 as time_t {
+                -1_i32 as time_t
             } else {
-                (((1 as i32 as time_t)
+                (((1_i32 as time_t)
                     << (::core::mem::size_of::<time_t>() as usize)
                         .wrapping_mul(CHAR_BIT as usize)
-                        .wrapping_sub(2 as usize))
+                        .wrapping_sub(2_usize))
                     - 1 as time_t)
                     * 2 as time_t
                     + 1 as time_t
@@ -265,7 +265,7 @@ pub unsafe fn ar_member_date(name: *const ::core::ffi::c_char) -> time_t {
     {
         val as time_t
     } else {
-        -(1 as i32) as time_t
+        -1_i32 as time_t
     }
 }
 /// # Safety
